pyproject.toml
[tool.poetry]
name = "quote-bot"
version = "0.1.0"
description = "Quotes constructor bot for civsoc"
authors = ["mihalin <stopa3.10.1996@mail.ru>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
aiogram = "^2.9.2"
python-dotenv = "^0.14.0"
Pillow = "^7.2.0"
aiohttp = "^3.6.2"
numpy = "^1.19.1"
scipy = "^1.5.1"
aiocache = "^0.11.1"

[tool.poetry.dev-dependencies]
isort = "^5.1.4"
pre-commit = "^2.6.0"
black = {version = "^19.10b0", allow-prereleases = true}

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
default_section = "THIRDPARTY"
known_first_party = "quote_bot"

[tool.black]
line-length = 120
exclude = '''
/(
    \.git
)/
'''

[tool.poetry.scripts]
bot = "quote_bot:main"
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

